more login/logout work
after successful login, show the game view. add a navbar to the page along
with a header that includes an account nav view with a logout link. require
a session for the game view. persist the session token in local storage.
disallow access to the app if the browser doesn't support local storage or
web sockets. end the session on logout.