add player signup
when signup is successful, initiate a session and show the game view. when
it fails, show errors inline.

add an alert subview to the game view so we can show alerts like "thanks for
signing up".

add better error handling for the api, returning a jsend error rather than
an ajax error.