handle world exits
clean up the session when the player logs out or quits. although, still
need to remove the character from the grid.

also, reuse an existing session if the the player is logged in and the
page is reloaded. in this case other players won't see duplicate "XYZ
logged in" messages.